fix(rust): migrate part of our lengths from usize to u32 (#1078)
We are using usize in a few places.
For all types interacting with memory allocations (vec, ...), that is
fine.

For the lenths we parse from varints? That seems buggy.

---------

+++ rust/mlt-core/src/errors.rs
@@ -44,7 +44,7 @@ pub enum MltError {
     #[error("found {0} bytes after the expected end of layer")]
     TrailingLayerData(usize),
     #[error("unexpected end of input (unable to take {0} bytes)")]
-    UnableToTake(usize),
+    UnableToTake(u32),
     #[error("unexpected stream type {0:?}")]
     UnexpectedStreamType(StreamType),
     #[error("unexpected stream type {0:?}, expected {1} for {2}")]
@@ -58,7 +58,7 @@ pub enum MltError {
 
     // Wire/codec decoding (bytes → primitives)
     #[error("buffer underflow: needed {0} bytes, but only {1} remain")]
-    BufferUnderflow(usize, usize),
+    BufferUnderflow(u32, usize),
     #[error("FastPFor decode failed: expected={0} got={1}")]
     FastPforDecode(usize, usize),
     #[error("invalid RLE run length (cannot convert to usize): value={0}")]
@@ -100,7 +100,7 @@ pub enum MltError {
     )]
     EncodingInstructionCountMismatch { input_len: usize, config_len: usize },
     #[error("struct child data streams expected exactly 1 value, got {0}")]
-    UnexpectedStructChildCount(usize),
+    UnexpectedStructChildCount(u32),
     #[error("unsupported physical encoding: {0}")]
     UnsupportedPhysicalEncoding(&'static str),
     #[error("unsupported physical encoding: {0:?} for {1}")]
@@ -108,7 +108,7 @@ pub enum MltError {
     #[error(
         "Extent {extent} cannot be encoded to morton due to morton allowing max. 16 bits, but {required_bits} would be required"
     )]
-    VertexMortonNotCompatibleWithExtent { extent: u64, required_bits: u32 },
+    VertexMortonNotCompatibleWithExtent { extent: u32, required_bits: u32 },
 
     // Geometry decode errors (field = variable name, geom_type for context)
     #[error("MVT error: {0}")]
